// ccm_react/src/client/models/BetaDistributionSpec.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A way to specify a beta distribution by giving alpha and beta.
 */
export type BetaDistributionSpec = {
  type: 'beta';
  distribution: 'beta';
  alpha: number;
  beta: number;
};