// ccm_react/src/client/models/CategoricalDistributionSpec.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CategoricalDistributionSpec = {
  type: 'categorical';
  distribution: 'categorical';
  /**
   * List of (probability, category) tuples
   */
  items: Array<any[]>;
};