// ccm_react/src/client/models/ConstantDistributionSpec.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A way to specify a constant distribution by giving a constant value.
 */
export type ConstantDistributionSpec = {
  type: 'constant';
  distribution: 'constant';
  value: number;
};