// ccm_react/src/client/models/UniformDistributionSpec.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A way to specify a uniform distribution by giving a range.
 */
export type UniformDistributionSpec = {
  type: 'uniform';
  distribution: 'uniform';
  range: any[];
};